+++ CHANGELOG.md
@@ -14,6 +14,7 @@ For information on changes in released versions of Teku, see the [releases page]
 
 ### Additions and Improvements
 
+- New databases will now default to `minimal` if no `--data-storage-mode` is explicitly set. Existing `prune` mode databases will need to explicitly set `--data-storage-mode=minimal` if they wish to use minimal data storage. This includes anyone not specifying `--data-storage-mode` in 23.6.2 or earlier.
 - Update attestation subnet subscriptions strategy according to [the spec changes](https://github.com/ethereum/consensus-specs/blob/dev/specs/phase0/p2p-interface.md#attestation-subnet-subscription). All nodes (including non-validating ones) will subscribe to 2 subnets regardless of the number of validators.
 - Added `/eth/v1/validator/{pubkey}/voluntary_exit` Validator API endpoint